refactor: split env helpers
